Should fix main problems
Was able to run build, lint, and test locally no errors. More render tests and split App.tsx reducing file size.
