--- star-ui/src/App.css
@@ -1,5 +1,0 @@
-/* App-level overrides -- intentionally minimal for tiling WM layout */
-#root {
-  width: 100%;
-  height: 100%;
-}
