+++ star-ui/src/test/setup.ts
@@ -1,1 +1,4 @@
 import '@testing-library/jest-dom';
+import { vi } from 'vitest';
+
+HTMLCanvasElement.prototype.getContext = vi.fn(() => null);
